V 1.09 - added Landscape mode. - set Footnote textsize same as webview textsize. - fixed : Duplicated shortcut on Go launcher

+++ Rigveda/RigVedaViewer/src/de/devmil/common/ui/color/HsvAlphaSelectorView.java
@@ -168,7 +168,7 @@ public class HsvAlphaSelectorView extends LinearLayout {
 		placeSelector();
 	}
 	
-	public int getAlpha()
+	public float getAlpha()
 	{
 		return alpha;
 	}
